third_party/WebKit/LayoutTests/external/wpt/css/css-tables-3/visibility-collapse-col-002-expected.txt

This is a testharness.js-based test.
PASS col visibility:collapse doesn't change table height
FAIL col visibility:collapse changes table widthassert_equals: col visibility:collapse changes table width expected 10 but got 222
Harness: the test ran to completion.
